Extract git/worktree code into src/lib/project/git.ts
All git-related code now lives in a single module under
`src/lib/project/git.ts`:

- From the deleted `src/lib/git.ts`: `parseGitHubRemoteUrl`,
  `getGitHubSlug`, `gitClone`, `gitPull`, `isWorkingTreeDirty`
- From `src/lib/project/refs.ts`: `normaliseGitRemote`,
  `detectGitWorktree`, `detectProjectWorktrees`, `retrieveGitRemote`,
  the `GitWorktree`/`ProjectWorktree` types, and the (now exported)
  `readGitInfo` + `extractGitHubSlug` helpers

`src/lib/project/refs.ts` is now just the ref derivation
(`projectRefs`, `projectSlug`, `projectId`) and consumes the git
helpers from its sibling module.

Tests have been combined into `src/lib/project/git.test.ts`
(`parseGitHubRemoteUrl`, `getGitHubSlug`, `normaliseGitRemote`,
`detectProjectWorktrees`), while `src/lib/project/refs.test.ts`
keeps the `projectRefs`/`projectSlug`/`projectId` suites.

Imports updated in `src/cli.ts`, `src/commands/system/update.ts`,
`src/lib/system/dotfiles.ts`, and `src/lib/project.ts`. The old
`src/lib/git.ts` and `src/lib/git.test.ts` are deleted.
