--- src/lib/git.ts
@@ -1,89 +1,0 @@
-import { execFile } from 'node:child_process'
-import { readFile } from 'node:fs/promises'
-import path from 'node:path'
-import { promisify } from 'node:util'
-
-import { runInherit } from './system/process.ts'
-
-const execFileAsync = promisify(execFile)
-
-/**
- * Parse a GitHub remote URL and extract owner/repo.
- * Supports:
- * - git@github.com:owner/repo.git
- * - https://github.com/owner/repo.git
- * - https://github.com/owner/repo
- */
-export const parseGitHubRemoteUrl = (url: string): string | null => {
-  // SSH format: git@github.com:owner/repo.git
-  // Owner and repo must not contain slashes, spaces, or other invalid characters
-  const sshMatch = url.match(
-    /^git@github\.com:([^/\s]+)\/([^/\s]+?)(?:\.git)?$/,
-  )
-  if (sshMatch) {
-    return `${sshMatch[1]}/${sshMatch[2]}`
-  }
-
-  // HTTPS format: https://github.com/owner/repo.git
-  // Owner and repo must not contain slashes, spaces, or other invalid characters
-  const httpsMatch = url.match(
-    /^https:\/\/github\.com\/([^/\s]+)\/([^/\s]+?)(?:\.git)?$/,
-  )
-  if (httpsMatch) {
-    return `${httpsMatch[1]}/${httpsMatch[2]}`
-  }
-
-  return null
-}
-
-/**
- * Get GitHub owner/repo from git remote, or null if not a GitHub repo.
- * Reads from .git/config to avoid spawning a subprocess.
- */
-export const getGitHubSlug = async (
-  projectPath: string,
-): Promise<string | null> => {
-  const gitConfigPath = path.join(projectPath, '.git', 'config')
-
-  try {
-    const gitConfig = await readFile(gitConfigPath, 'utf-8')
-
-    // Check remotes in priority order: origin first, then github
-    for (const remoteName of ['origin', 'github']) {
-      const remoteMatch = gitConfig.match(
-        new RegExp(`\\[remote "${remoteName}"\\][^[]*url\\s*=\\s*([^\\s\\n]+)`),
-      )
-
-      if (remoteMatch) {
-        const slug = parseGitHubRemoteUrl(remoteMatch[1])
-        if (slug) return slug
-      }
-    }
-
-    return null
-  } catch {
-    return null
-  }
-}
-
-/** Clone a git repository into the target directory. */
-export const gitClone = (url: string, target: string): Promise<boolean> => {
-  return runInherit('git', ['clone', url, target])
-}
-
-/** Run `git pull` in the given directory. */
-export const gitPull = (cwd: string): Promise<boolean> => {
-  return runInherit('git', ['pull'], { cwd })
-}
-
-/** Check if a git working tree has uncommitted changes. */
-export const isWorkingTreeDirty = async (cwd: string): Promise<boolean> => {
-  try {
-    const { stdout } = await execFileAsync('git', ['status', '--porcelain'], {
-      cwd,
-    })
-    return stdout.trim().length > 0
-  } catch {
-    return false
-  }
-}
